+++ packages/article-summary/fixtures/article-multi.js
@@ -36,7 +36,7 @@ export default ({
       }
     ]
   },
-  content: () => (
+  content: (
     <ArticleSummaryContent
       ast={[
         {
@@ -72,5 +72,5 @@ export default ({
     date: "2017-11-17T00:01:00.000Z",
     publication: "TIMES"
   },
-  headline: () => <ArticleSummaryHeadline headline={headline} />
+  headline: <ArticleSummaryHeadline headline={headline} />
 });