+++ packages/article-error/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.78](https://github.com/newsuk/times-components/compare/@times-components/article-error@0.2.77...@times-components/article-error@0.2.78) (2019-05-20)
+
+**Note:** Version bump only for package @times-components/article-error
+
+
+
+
+
 ## [0.2.77](https://github.com/newsuk/times-components/compare/@times-components/article-error@0.2.76...@times-components/article-error@0.2.77) (2019-05-20)
 
 **Note:** Version bump only for package @times-components/article-error