+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.6](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.2.5...@times-components/article-in-depth@3.2.6) (2019-05-21)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.2.5](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.2.4...@times-components/article-in-depth@3.2.5) (2019-05-20)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.2.4](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.2.3...@times-components/article-in-depth@3.2.4) (2019-05-20)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.2.3](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.2.2...@times-components/article-in-depth@3.2.3) (2019-05-20)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
 ## [3.2.2](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.2.1...@times-components/article-in-depth@3.2.2) (2019-05-20)
 
 **Note:** Version bump only for package @times-components/article-in-depth